:white_check_mark: add Unit test for CRUD API actions for tasks entity

+++ api/task.go
@@ -1,7 +1,9 @@
 package api
 
 import (
+	"encoding/json"
 	"fmt"
+	"log"
 	"net/http"
 
 	"github.com/gin-gonic/gin"
@@ -9,7 +11,7 @@ import (
 )
 
 func postTask(c *gin.Context) {
-
+	loggeur := log.Default()
 	var newTask model.Task
 
 	if err := c.BindJSON(&newTask); err != nil {
@@ -17,17 +19,26 @@ func postTask(c *gin.Context) {
 		return
 	}
 
-	newTask.CreateTask()
+	err := newTask.CreateTask()
+
+	if err != nil {
+		loggeur.Printf("an error occured during the insertion.")
+		msg, _ := json.Marshal(err)
+		c.AbortWithStatusJSON(http.StatusInternalServerError, msg)
+		return
+	}
+
+	loggeur.Printf("everything went well.")
 	c.IndentedJSON(http.StatusCreated, newTask)
 }
 
 func returnTask(c *gin.Context) {
 	id := c.Param("id")
 
-	task, err := model.ReturnTask(id)
+	task := model.ReturnTask(id)
 
-	if err != nil {
-		c.Error(err)
+	if (task == model.Task{}) {
+		c.AbortWithStatus(http.StatusNotFound)
 	}
 
 	c.IndentedJSON(http.StatusOK, task)
@@ -38,7 +49,7 @@ func returnTasks(c *gin.Context) {
 	tasks, err := model.ReturnTasks()
 
 	if err != nil {
-		c.Error(err)
+		c.AbortWithError(http.StatusInternalServerError, err)
 	}
 
 	c.IndentedJSON(http.StatusOK, tasks)
@@ -48,23 +59,27 @@ func updateTask(c *gin.Context) {
 	var updatedTask model.Task
 
 	if err := c.BindJSON(&updatedTask); err != nil {
-		fmt.Println(err)
+		c.AbortWithError(http.StatusInternalServerError, err)
 		return
 	}
 
-	updatedTask.UpdateTask()
+	err := updatedTask.UpdateTask()
+	if err != nil {
+		c.AbortWithError(http.StatusInternalServerError, err)
+	}
+
 	c.IndentedJSON(http.StatusOK, updatedTask)
 }
 
 func deleteTask(c *gin.Context) {
 	id := c.Param("id")
 
-	taskToDelete, err := model.ReturnTask(id)
+	taskToDelete := model.ReturnTask(id)
 
-	if err != nil {
-		c.Error(err)
+	if (taskToDelete == model.Task{}) {
+		c.AbortWithStatus(http.StatusNotFound)
+	} else {
+		taskToDelete.DeleteTask()
+		c.IndentedJSON(http.StatusOK, taskToDelete)
 	}
-
-	taskToDelete.DeleteTask()
-	c.IndentedJSON(http.StatusOK, taskToDelete)
 }
